requirements.txt
Django>=5.2.1
djangorestframework>=3.15.2
djangorestframework-simplejwt>=5.4.0
Pillow>=9.0
psycopg2-binary>=2.9
gunicorn>=20.1
